pyproject.toml
[project]
name = "example_dspy_redis"
version = "0.1.0"
description = "A template project using Python, uv, Docker and GitHub Actions"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dspy>=3.1.2",
    "python-dotenv>=1",
    "redis>=7.1.0",
]

[project.scripts]
main = "example_dspy_redis.app:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
